+++ appmgr/src/backup/restore.rs
@@ -192,13 +192,13 @@ pub async fn recover_full_embassy(
     .await?;
     secret_store.close().await;
 
-    let rpc_ctx = RpcContext::init(ctx.config_path.as_ref(), disk_guid).await?;
-    let mut db = rpc_ctx.db.handle();
-
     Ok((
         os_backup.tor_key.public().get_onion_address(),
         os_backup.root_ca_cert,
         async move {
+            let rpc_ctx = RpcContext::init(ctx.config_path.as_ref(), disk_guid).await?;
+            let mut db = rpc_ctx.db.handle();
+            
             let ids = backup_guard
             .metadata
             .package_backups
